rename: userPreference from timecardUserPreference
